Redo AdmissionsTable, Admission data model.
