+++ database/seeders/DatabaseSeeder.php
@@ -43,7 +43,7 @@ class DatabaseSeeder extends Seeder
         $this->call(\Database\Seeders\SalesDistributorSeeder::class);
         $this->call(\Database\Seeders\DocumentSeeder::class);
         
+        $this->call(\Database\Seeders\AdmissionsTableSeeder::class);
         $this->call(\Database\Seeders\AdmissionSeeder::class);
-        $this->call(\Database\Seeders\AdmissionFilmSeeder::class);
     }
 }
